Update package-lock.json with new dependencies and versions
The package-lock.json file has been updated with new dependencies including 'function-bind', '@ljharb/through' and others. Additionally, the versions of several dependencies such as 'chalk', 'cli-width', 'tslib', and more have been updated. Changes are made to ensure the functionality of the app with the integration of new packages and their updated versions.